freedreno/kgsl: avoid bo memory leak
No shared GPU memory at all was getting freed because of DRM kgsl not
using the kernel interface properly. There is a "gsl_memdesc_t" that is
filled in on allocation and supposed to be passed as-is to the kernel
interface on freeing, but drm only fills in the gpuaddr field (which is
regarded as 32-bit handle throughout the interface).

Add a hash table on the kgsl device to keep track of the memdesc
associated with a handle.